refactor: rename ProcessSteps component to Status
